Prover validate boundary and transition constraints

+++ math/src/circle/cosets.rs
@@ -57,7 +57,7 @@ impl Coset {
     pub fn get_coset_points(coset: &Self) -> Vec<CirclePoint<Mersenne31Field>> {
         // g_n the generator of the subgroup of order n.
         let generator_n = CirclePoint::get_generator_of_subgroup(coset.log_2_size);
-        let size: u8 = 1 << coset.log_2_size;
+        let size: usize = 1 << coset.log_2_size;
         core::iter::successors(Some(coset.shift.clone()), move |prev| {
             Some(prev + &generator_n)
         })
